support mistral too
